+++ src/ocp/awesome_mapper2/functions-am3-process.go
@@ -193,12 +193,12 @@ func init() {
 				&domain.UpdateRedfishResourceProperties2{
 					ID: mp.uuid,
 					Properties: map[string]interface{}{
-						"Oem/EnergyConsumptionStartTime": epoch2Date(data.CwStartTime),
+						"Oem/EnergyConsumptionStartTime": epoch2Date(int64(data.CwStartTime)),
 						"Oem/EnergyConsumptionkWh":       int(data.CumulativeWatts / 1000),
 						"Oem/MaxPeakWatts":               data.PeakWatts,
-						"Oem/MaxPeakWattsTime":           epoch2Date(data.PwReadingTime),
+						"Oem/MaxPeakWattsTime":           epoch2Date(int64(data.PwReadingTime)),
 						"Oem/MinPeakWatts":               data.MinWatts,
-						"Oem/MinPeakWattsTime":           epoch2Date(data.MinwReadingTime),
+						"Oem/MinPeakWattsTime":           epoch2Date(int64(data.MinwReadingTime)),
 						"Oem/PeakHeadroomWatts":          data.PeakHeadRoom,
 						"PowerConsumedWatts":             data.InstWattsPSU1_2,
 						"PowerAvailableWatts":            data.PeakHeadRoom,
